mysite/requirements.txt
asgiref==3.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.0.4
django-ckeditor==6.6.1
django-embed-video==1.4.9
django-heroku==0.3.1
django-js-asset==2.1.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.6
python-decouple==3.6
pytz==2022.4
requests==2.31.0
sqlparse==0.4.3
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==5.2.0
django-admin-thumbnails==0.2.8

